Use pre-resolved requirements for examples
This adds a resolver script which is used to pre-resolve requirements
(existing always as wheels) into a `requirements.bzl`.

It's currently not the most robust: different dependency sets can
trample over each other because we don't include the name in the
generated repository names.

That said, we do use this in production for Python 2.7 and 3.6, building
docker images from both macOS and Linux machines.